+++ doc/source/changelog.md
@@ -20,6 +20,7 @@ Primarily developped with Godot 4.2.
 - `VoxelBuffer`: exposed `fill_area_f`
 - `VoxelEngine`: added methods to get the version of the voxel engine
 - `VoxelGeneratorGraph`: Added GPU support for the `Select` node
+- `VoxelStreamSQLite`: Added support for `user://` paths (via internal call to `ProjectSettings.globalize_path()`)
 - `VoxelTool`:
     - Added `grow_sphere` as alternate way to progressively grow or shrink matter in a spherical region with smooth voxels (thanks to Piratux)
     - `do_box` with smooth voxels now uses a proper box SDF, to improve quality. Before it was a solid fill, which could cause artifacts